Feature/os

+++ src/client/src/tscLocalMerge.c
@@ -914,8 +914,8 @@ static void doFillResult(SSqlObj *pSql, SLocalReducer *pLocalReducer, bool doneO
 
     if (pQueryInfo->limit.limit >= 0 && pRes->numOfRowsGroup > pQueryInfo->limit.limit) {
       /* impose the limitation of output rows on the final result */
-      int32_t prevSize = pFinalDataPage->num;
-      int32_t overflow = pRes->numOfRowsGroup - pQueryInfo->limit.limit;
+      int32_t prevSize = (int32_t)pFinalDataPage->num;
+      int32_t overflow = (int32_t)(pRes->numOfRowsGroup - pQueryInfo->limit.limit);
       assert(overflow < pRes->numOfRows);
 
       pRes->numOfRowsGroup = pQueryInfo->limit.limit;
@@ -984,7 +984,7 @@ static void doFillResult(SSqlObj *pSql, SLocalReducer *pLocalReducer, bool doneO
 
   if (pRes->numOfRows > 0) {
     if (pQueryInfo->limit.limit >= 0 && pRes->numOfRows > pQueryInfo->limit.limit) {
-      int32_t overflow = pRes->numOfRows - pQueryInfo->limit.limit;
+      int32_t overflow = (int32_t)(pRes->numOfRows - pQueryInfo->limit.limit);
       pRes->numOfRows -= overflow;
       pFinalDataPage->num -= overflow;
